Beta working version
Now I need to refactor, test for dumb things.
But in general I think that it is working.

Things that I imagine that could be problem is the use
of logged and absent, to easy, instead I should be eitheir
using a hard to guess token or better yet a random token
for each user. But this would be much harder to implement

Signed-off-by: luatil <luasantilli@gmail.com>